#39 rename "storage" to "data-store"
